Initial commit.
Succint is a jQuery plugin to provide a user with feedback
as to how many characters they have remaining when completing a
text-type HTML form field. It will also prevent the user from
entering any more text into the field when the limit has been
reached.

Included jquery 1.3.2, as that's the current production version
I'm working with.